requirements.txt
requests===2.24.0
lxml===4.5.1
icalendar==5.0.7
python-dateutil==2.8.2
python-decouple==3.8